Move ADT related code to a sub module for type info
